Get rid of compiler warnings

+++ src/EngineCore/Dx11/ParticlesRenderPass.hpp
@@ -80,7 +80,7 @@ namespace EngineCore {
                                 unsigned int particles_cnt = 0;
                                 if (!particles_indices.empty()) {
                                     particles_rsrc = particles_mngr.getComponent(particles_indices.front()).particle_data;
-                                    particles_cnt = particles_mngr.getComponent(particles_indices.front()).particle_count;
+                                    particles_cnt = static_cast<unsigned int>(particles_mngr.getComponent(particles_indices.front()).particle_count);
                                 }
 
                                 data.particles_rtd.push_back({ rt.shader_prgm, particles_rsrc,particles_cnt });
@@ -102,7 +102,10 @@ namespace EngineCore {
                             constantBufferData.pSysMem = data.particles_cbs.data();
                             constantBufferData.SysMemPitch = 0;
                             constantBufferData.SysMemSlicePitch = 0;
-                            const CD3D11_BUFFER_DESC constantBufferDesc(sizeof(Data::ParticlesConstantBuffer) * data.particles_cbs.size(), D3D11_BIND_CONSTANT_BUFFER);
+                            const CD3D11_BUFFER_DESC constantBufferDesc(
+                                static_cast<UINT>(sizeof(Data::ParticlesConstantBuffer) * data.particles_cbs.size()),
+                                D3D11_BIND_CONSTANT_BUFFER
+                            );
                             winrt::check_hresult(
                                 resources.d3d11_device->CreateBuffer(
                                     &constantBufferDesc,
@@ -217,7 +220,7 @@ namespace EngineCore {
                                     //TODO
                                     context->DrawInstanced(
                                         resources.particles_resources[rt_idx].particles_cnt * 6, // Vertex count
-                                        frame.m_view_projections.size(),                         // Instance count
+                                        static_cast<UINT>(frame.m_view_projections.size()),      // Instance count
                                         0,                                                       // Base vertex location
                                         0                                                        // Start instance location
                                     );
